Cp 13: Multi-game project
There are 2 games in the project now. Miner and MyCosmic. MyCosmic is inspired by a doom mod called MyHouse.wad. I work on mycosmic when I am frustrated with the Miner for how difficult it is. The idea of making 2 games in a project is... less dependency management, smaller project size, and faster iterative testing. The con is diamond dependency. If you are interested in what I planned for chunk threading? Look at the design at the bottom. Note there's no chunk unloading yet.