Cp 14: Chunk culling height bug fix
It works with non-cube chunks now. But not with rectangle chunks. Reduced project\model\kira\import\material_atlas_34048_1.res by removing the image and linking it again with exported one. Reimported the whole project to stop the "missing UID" messages littering around the console. Made a batch file to get rid of the .godot folder for trimming project size to upload.