Cp 16: Rectangle culling bug fix
Now it supports rectangle chunks! But there were index bugs when Kira mined the block outside the chunk. It will crash the game. A new bug discovered! And it's challenging to detect due to Godot supporting the awful negative index. And I don't know how to turn on the warning for the negative index or disable it. Sometimes it's out of index. This bug is much easier to solve with a 3d array. But it's not easy to make 3d arrays in Godot. So I sticking with the 1d arrays.