Cp 21: Using RenderingServer to make the chunks
I had 3 options, from easiest and unportable to hardest and most portable.
The first one is to stick with the custom Godot engine. if I did it, that means I had to maintain it. I don't want to maintain a black box. The Gridmap is constantly changing in the source code, it breaks with every update(4.0.2 between 4.0.4 rc1). Unless I apply the commits from the Godot repository.
The second option is... Stick with multiple GridMap, which is a nightmare to fix the bugs. As you constantly try to figure out detecting where Kira is breaking the block. And I saw the source code (grid_map.cpp)[https://github.com/godotengine/godot/blob/master/modules/gridmap/grid_map.cpp], it uses hashmap in vector as type. Which is overengineered for my use case.
The third option is, to use RenderingServer. It's the hardest one, but it will make sure I know what the Godot engine is doing. So I can maintain it much easier if Godot 5 breaks it. And it's fully GDScript. So it's very portable for the engine.