Cp 25: Using triangle strip to create cube
The triangle strip is way more efficient than triangle. because it uses 1 index instead of 3 after the 2nd index. Even if it's not any more efficient than the triangle, it takes up 8 bytes instead of 24 bytes per square in the RAM. I am testing this on MeshInstance3D first, then move to RenderingServer once it's ready.