Cp 29: ditched greedy meshing, using QOI & xxh32
i realized greedy meshing scale performance gain in linear instead of exponential. so back to texture atlas again. and separating the mesh cost so much CPU cycles for just to be greedy meshing compatible. it's not even worth it to consider. i am going to use LOD chunks instead.
then... you did saw the 2 extra file. xxh32.hpp and qoi.gdextension. i am planning to store the 8**3 chunks in the 4096x4096 qoi image. the chunk is shockingly smooth as it's generated using simplex, similar to photograph. that makes the chunks perfectly for storing in an image format. but i need a fast one, or people will get mad at me for storing their chunks for 10 minutes. the Red channel will be storing the block ids from the chunk. Blue and Alpha channel will be reserved for mods. whatever used as the extension of block id or something. Green channel will be used for hashes from xxh32 or mod settings.
xxh32 will be used for 16-bit block-level deduplication instead of 32-bit.