Cp 30: Moved to RenderingServer
The lighting and the physics doesn't work yet. I was trying to append the generated collision from the MeshInstance3D to PhysicsServer3D to make it work. I still have no idea how to fix the lighting from the chunk or what's the problem with it, it might be incorrect vertices problem at last time I checked. And the chunk takes forever to generate for 256 cubed. I didn't go straight for infinite generation. I'm going for old generation like back in day Minecraft is called Minecraft PE instead of Minecraft bedrock edition. Until the chunk generating speed is improved, I won't add infinite generation. I remembered the MCPE videos are in 0.3.2. That's when 2012. Now it's 2023, you know.